"7:30pm. I am up. I feel somewhat refreshed from that nap.
I've been thinking how I want to approach the pool scene and it seems to me that finding a way to assign shading layers and then just working on those is the way to go. It would be too much work to translate the scene in Solaris. Solaris is something I'd use for a new scene, but I do not feel like redoing all that layouting from scratch.

I got a reply from Stormz saying he does not know. I'll have to ask on odforce.

https://forums.odforce.net/topic/50600-how-do-i-export-shading-groups-in-usd-files/

I asked on odforce. Somebody here should know how to do it.

7:45pm. Let me watch some USD tutorials.

https://youtu.be/H-RYtuacUlM
Solaris Workshop - Intro to USD Concepts | Rob Stauffer | Houdini HIVE Worldwide

Let me watch this. One of the tutorials is 4h and I do not feel like watching that. It seems to be composed of 1h talks, but who knows where the information I need is.

7:45pm. I should have mentioned it, but it is also annoying that when I exported the USD I got various weird errors in the node. The export itself seems to have gone through fine though.

8pm. I ended up sufing /a/ and just skimming the thing in a minute. This stuff is so boring and bureaucratic. From my 1m skim it never goes into exports.

https://youtu.be/DI09YJLZ_6k?t=72
Solaris Workshop - Asset Prep for Solaris | Jeff Wagner | Houdini HIVE Worldwide

This seems more promising. It actually mentions shop_materialpath.

> What is Clarisse PLE?

> Our Clarisse Personal Learning Edition (PLE) gives you the opportunity to learn and explore all aspects of Clarisse iFX and Clarisse BUiLDER at your own pace and for free. For non-commercial use only, the Clarisse PLE gives you access to all the features* available in commercial versions of Clarisse, and lets you save up to 3 watermark-free renders (2560x1440 max) every 24 hours. As long as you’re using your renders for non-commercial purposes exclusively.

I'll consider Clarisse PLE in order to get Angie at some point, but not now. This restriction of 3 renders a day is not severe for me.

https://youtu.be/DI09YJLZ_6k?t=74
> Focus on getting geometry data out of SOPs and into LOPs converted to USD.

This seems like the talk I was looking for.

https://youtu.be/DI09YJLZ_6k?t=178

> name + path attributes. Maybe the path attribute is what I need for shading groups?

It seems the alembic files have shading groups. Let me import them in Houdini.

8:20pm. They don't have attributes. ...Actually they do, they are just several layers deep. How do I unpack an alembic?

Using the unpack SOP. Now I can see what is underneath. And that would be a ton of primitive groups. Those don't work for me in USD. Maybe I have to pack the object first? Who knows.

Let me go back to the tutorial.

https://youtu.be/DI09YJLZ_6k?t=399

It outright says here that prim groups are for material assignment.

8:35pm. I can't figure it out. Given that alembic uses prim groups for shading group assignment, and the above slide, it is almost certain that shading groups are set through primitive groups. But nothing is coming up. Maybe I have to prefix them with material_?

...No, what if I instead set the subset groups in import?

I got it!

I just needed to set the subset groups!

Now the shading layers show up. Still, this is more complicated than I thought. I'd have many primitive groups without this. Also the material path has been set on packed primitives for the sake of V-Ray. Just how will I translate that?

Well, on the plus side, all materials paths have S_ prefix. So it might be possible to automate this. On the other...

8:50pm. Actually could I put it on top of a pack? Let me try it out.

///

I see now that I can use primitive groups to set shading groups in USD. To make this work, I have to mark them as Subset Groups in Import Data in the USD Export nodes either LOP or SOP.

I've also figured out that putting in * works for exporting groups even in packed and nested packed primitives, but that would be too much. Instead I should be able to prefix the groups and then put in something like s_* in Subset Groups to properly export them.

///

Posted this in that forum. This is the solution I am going to aim for. It should work well.

> Now that I got this out of the way, why is Clarisse complaining about the asset being left-handed? How do I export it as a right handed asset instead?

I also need to figure this out, but it is not a big deal.

https://www.sidefx.com/docs/unreal/_coordinate_system.html

It has something to do with coordinate systems. It makes sense to talk handedness here.

https://youtu.be/DI09YJLZ_6k?t=581

Let me just watch this video. Google put up this search. The video itself will probably cover it.

https://youtu.be/DI09YJLZ_6k?t=980

No, I can't be assed to watch this. It is too boring. I have everything I need. I do not need to mess with conversions. So what if it take a bit extra to load?

With this I should be able to start work on the scene, but I want to deal with the course I got first. This won't be hard.

Before I close let me organize and install Nuke.

...I'll leave installing Nuke for later, if at all. It looks tedious to crack. I'll watch course first before I decide whether I want a compositing software.

9:35pm. So far the Houdini parts of the course look quite simple, so I am not really motivated to try and replicate it, but I want to watch the Clarisse parts to see what he does.

Yeah, I should compromise. Let me just watch the course and I'll do it myself if I feel like it.

What is really important to me is that I work on the pool scene and finish it.

9:40pm. Let me close here. I'll watch some anime and go to bed."