"10:40am. I am up. Let me chill for a while and then I will start by watching the procedural texturing tutorial. Did I get a reply to any of the Clarisse questions? No. This is the first time I've experienced something like this. Are they really holding me back from asking pertinent questions on their forum? At this point I can assume that what I want from those questions will be resolved negatively.
11:25am. Let me start. I've listened to the XPirates Geoscape OST a few times, and my conclussion is that it is very difficult to compete with the original Geoscape theme from UFO. It is easily one of the best musical compositions of all time.

A lot what is going on in this era is not appreciated properly in the present. After the Singularity hits, it is very likely that people will look back at this period as a mythical era.

Games themselves are yet to take their proper role.

https://youtu.be/9ngbOV0YKbk
Advanced Procedural Texturing

Let me start by watching this. The first time around I was too unfamiliar with Clarisse to understand it, but now I'll be able to make progress.

https://youtu.be/9ngbOV0YKbk?t=76

Ah, this is a way to get the UVs.

Lately I really feel like I am going senile. So many thoughts are going through my mind...

Ah, yes. A question I wanted to ask was: How did you get Houdini to show a whole city without it shitting itself?

There was a clip in both the keynote and the USD exports tab. I'll leave that for later.

https://youtu.be/9ngbOV0YKbk?t=79

Also, watching this makes me realize how he got those quadrants in blender. The black square is the negative position, and the rest go sharply positive quickly.

https://youtu.be/9ngbOV0YKbk?t=136

What is the difference between randomizing a texture and creating a random one?

https://youtu.be/9ngbOV0YKbk?t=132

Why did he get squares like this. When I tried it yesterday I just got a white noise texture.

https://youtu.be/9ngbOV0YKbk?t=203

Ah, the color picked can show values beyond 1. It did not occur to me that this could be a use of the pixel inspection.

https://youtu.be/9ngbOV0YKbk?t=239

Let me pause here. I want to understand what randomize is doing. How did he get those squares?

https://youtu.be/9ngbOV0YKbk?t=55

I do not get it. Why did changing the projection cause it to tile?

12pm. Hmmm, I see. It seems randomize is a noise function that takes in color as input. That means the same color gets mapped to the same thing.

12:15pm. https://youtu.be/9ngbOV0YKbk?t=360

I can follow this video perfectly now.

https://youtu.be/9ngbOV0YKbk?t=766

Sigh, why don't the remap nodes start with grayscale by default. Why just one point white? Unlike in Houdini there are no presets either.

https://youtu.be/9ngbOV0YKbk?t=1064

It seems there is a specialize point cloud texture.

12:55pm. https://youtu.be/R3usfNNH8wg
Clarisse Workflows: Creating Procedural Dunes

Let me watch this. There is a desert scene as well in the intro peom.

https://youtu.be/R3usfNNH8wg?t=233

Hmmmm, hmmm, what if I started with a voronoi texture, inverted the colors and then squished it. But I am not sure of being able to manage the rotation.

https://youtu.be/R3usfNNH8wg?t=544

Ah, it is possible to use an expression like this.

1:15pm. I learned quite a bit. When it comes to making deserts I'll definitely make use of this knowledge. You can really get a lot of power through procedural texturing.

I really meant to use sculpting for a lot more things, but now that I am learning different techniques, I am leaning more and more towards the programming approach.

1:20pm. It might seem that I haven't moved much from months ago, but previously I definitely was not paying attention to texturing. Right now I feel like I've gotten my modeling concerns out of the way which allwos me on to focus a lot more on the 'painterly' aspects of 3d art. After I internalize this, I will have completely conquered 3d illustration.

Let me have breakfast here. I am overdue."