"10:10am. I am up. Let me chill a bit and I will start.
10:30am. Let me start. If I start reading Yurushia's story I'll waste the whole morning. I slept well tonight, and am feeling rested right now. This is the opposite of yesterday. So I should do some modeling. Oh yeah, let me set the course to download.

10:35am. Let me drop the clipping mask. What I did with the floor was good. Even if the tiles themselves are repetitive, by slapping a non-repetitive pattern of scratches on top, you can get something better.

10:45am. Let me try a gradient ramp. I just can't the overall colors to match my own laminate. I tried reducing the contrast and increasing the saturation, but that still just makes it look bleached out.

Should there be something like the curves too for rgb. Blender had a thing where I could change the contrast directly using curves. I neeed something like that here.

> Error. Captcha expired. Try again in 1 hour.

Forgot to press download on RapidGator and got this. They are just messing with mehere here. No, I won't buy your shitty account.

10:55am. There is the remap node, but I can't figure out how to add extra points on the curve. Sloppy work by Isotropix.

Let me try the gradient ramp.

https://en.wikipedia.org/wiki/Shades_of_brown

I really need once and for all figure out what the standard brown is.

11:15am. I give up. I can't the floor to have the right feel no matter how much I mess with the contrast, saturation or ramps, The overall texture itself is not right, so messing with the globals just gives me the same pattern in the end.

Just forget about it.

I need to move on to the props.

Focus me, just load up Houdini and make the thing.

11:30am. Mhhhh, I knew this would be hard. Let me restart.

Also it seems that mirroing while doing the booleans is a bad idea. When I try to boolean an open mesh, I get outright cutting.

12:55pm. This sucks so bad. Trying to model this in Houdini is sucking the life out of me. For an object of this complexity it is not like I can afford to make it generic either.

I thought it would be easier in Houdini, but I should try to make in Blender as well for comparison.

1pm. The UI matters a great deal here. It is just so awkward to do this in Houdini

Let me stop here. Since this is my first time doing serious independent modeling, of course I am not going to be taking off like a rocket. Something like this is the best possible exercise I could make to level up my 3d skills at my current level of development.

I watched a lot of things, but I am severely lacking in practice. I need to do this smartly and refine my workflow. It won't take me too long to break into 3/5 if I can clear these hurdles.

2:40pm. Done with breakfast and chores. I really don't feel like it, a part of me just wants to drop all of this laze around around for the rest of my days. At this point, thanks to my accumulated programming ability, the easy route for me is definitely to take up programming for money. No need to learn new things or struggle that way.

But the reason why I got good at programming in the first place is due to endless struggle. It was not by taking the easy way.

Something like making a desk is not something that should be causing me consternation. A challenge like this should be done in under an hour easy for a skilled artist.

Obviously, Houdini is not what I should be using here. I can't be both quick and generic. I'd be much better served of being able to make quick prototypes. For texturing I should learn how to draw masks on which to later slap noise on.

Now focus me. Let me give it a try from start.

4:20pm. That was engrossing. I only have that cylinder at the bottom which acts as a leg placer.

I am looking at the objects and I see that in 3.1 Blender has Nurbs surfaces like Houdini. I haven't tried them out, but I am impressed by its pace of development. It really feels nice to work in a program that is as stable and snappy as it.

4:35pm. The bisect tool is really nice.

Now I'll be able to mirror the whole mesh cleanly. But...wasn't there an explicit symmetrize option?

4:50pm. Yeah, it works great. I should not have done that bisect as it left a cut in the middle. But it was easy to remove. Let me do the finishing touches.

Let me make the vertex groups for the bevels.

Ah, right. I can just use weights. For those which I want less bevel, I can just lower the strength.

5:55pm. Had to take a lunch break. The creases are giving me trouble. What is happening that other creases influence the maximum offset and there is a gap in one of the boards that is really close to the outer edge.

Right now I am exploring the ways to fine tune it properly. Maybe if I give the edges low weights it will work.

Honestly, it is cool that Blender has actual edge attributes. Houdini has points, primitives and vertices, but no edges which is annoying.

6pm. Focus me. I need to keep going forward. I am going to deal with this desk today. The modeling part is almost done, I just need to figure the beeveling and then texture it.

This last part is admittedly not easy as the top has some very unique wear patterns. I am not sure if it is worth to try to replicate it, if I do, it will probably take me a day or two of experimenting. But it could be worth it. I am studying texturing at this point in time. Now that I've established my basic modeling workflow, this pattern is probably the hardest part of the entire room.

Yeah, I really should replicate it. I mean, I myself am 12h a day in front of the computer, so it fairly important part of it as it is directly between me and the monitor.

6:50pm. I should have been more wary of ngons. Next time I'll avoid using them. I decided to redo the piece I was working on. I got quite into it.

Now let me finish beveling the rest of the piece.

7:15pm. Trying to remove the middle seam that happens due to symmetrizing can be troublesome. Just now disolve edges left vertices behind and ended up deleting faces for one object. Next time I'll just leave the seam there, it is not worth bothering with.

What I am looking at is pretty much my own desk. It is a really convincing 3d model. I do not know whether I messed it up somewhere, but it feels right when I look at it, so that if fine. Even if I got something wrong compared to the real life one, it is not like anybody will be raiding my room to check so who cares.

Yeah, this is the way to do it. I am going to have to assign a few shading groups to different parts and then export it as obj. Or maybe USD would be better, let me go with that. I have no reason to favor obj.

7:25pm. Right now I am thinking how I could make the texture. At first glance it reminded me of polynomials, but instead of that, why not make my reasoning's starting point a topological chart instead.

What I could do is make some fine anti aliased noise, quantize, edge detect. That will give me the eshes that I want. I could thicken the edges a bit, multiply that with some rougher noise and voila, I have my mask.

Alternatively, I do not even need quantization and edge detection. I could just use a ramp and whipsaw it between black and white to get the edges. I did the same thing for the ceiling patterns. I just need to push the idea further this time.

Yeah, that would be really good. I should use constant interpolations this time. The best part of this is that I could do it in Clarisse as well. No need to mess with CHOPs.

7:40pm. Let me close here. I'll put the plan into action tomorrow. Right now I don't feel like it anymore. A day's work is a day's work.

Tomorrow I am going to do the UV unwrapping and then texturing in Clarisse. After that I'll deal with the rest of the props.

When I've completed the room, I'll have achieved something significant."