"10:15pm. Actually, let me do a bit more. I had some hints.
10:50pm. I figured it out completely now. The reason for the seams was definitely insufficient padding between the islands. I also had the problem of with post-beveled UVs stretching past the boundaries. This wasn't obvious when I was looking at the UVs in Blender since the beveling was done using a modifier, but I noticed it immediately in Substance Painter's UV view.

I still need to experiment a bit to figure out how much I can trim the padding, previously when I tried it I put in too small of a value and did not manage fix the seam.

11:25pm. It seems that setting the padding to when packing or unwrapping 0.01 is fine.

Ok, I am done for the day. This issue was literally me stepping on a mine. I should be able to make better progress tomorrow."