"4pm. The thunder did not last much, it went away as soon as I went to bed. It gave me time to think. I want to do more playing around with UVs. I need to go back to that metal piece and see whether setting the face area to 1 improves the seam placement (Edit: Yes it does). Right now it is extremely awful. After that, I want to see if setting the number of segments to even in bevel preserves the seams. It is actually a lot easier to do UVing on unbeveled surfaces.
4:10pm. How annoying. The bevels move the seams utwards rather than leave it in the middle. Well, no matter.

4:40pm. If you use select shortest path, it is possible to auto tag seams with it. I set the shortcut to something easy to reach like \ . In that case it is only slightly harder to use than unbeveled surfaces. When I set the face area to 1, the auto UV is decent on this example. It is not better than manual, but it is workable.

Ok...

4:40pm. I hate these small scans for Crow Record. Ok, let me take a break to read the second chapter and I will do some actual work.

4:50pm. Ok, enough of UVs, I am sick of them. I happy with my current grasp of them. It is not such a difficult subject.

Let me focus on texturing.

Time to fire up Substance.

https://feedback.substance3d.adobe.com/forums/261284-substance-3d-painter/suggestions/6626915-swap-foreground-and-background-color-with-shortcut

First let me resolve this problem. Why doesn't pressing X work for me?

It works with the mask layer, that switches it from black to white, but what about the paint layer? Why doesn't that work?

https://substance3d.adobe.com/documentation/spdoc/smudge-tool-143720512.html
> The simplest way to use the smudge tool is to use it directly on the content of a painting layer, as a regular paint tool.

> A smarter way to use the smudge tool is to create a painting layer and set all the channels of the layer to the "Pass through" blending mode. This will allow to smudge in a non-destructive way over all the layers located below the "smudge layer". The layers below remain intact and all the modifications applied later will be taken in account by the smudge layer.

Good advice.

Right now I am looking at the docs. I'll have to go through the docs for Clarisse as well at some point. All in due time.

For some reason the image links are broken in the docs.

5:20pm. I am looking at the shortcuts, and X is the invert grayscale tool.

There aren't too many to deal with. Ok...

https://youtu.be/H4rgnscD_3U
Passthrough Blend Mode in Substance Painter

I won't rush myself. Let me immerse myself in study. UVs I have down. Now there is just this and I'll be able to be an effective 3d artist.

I remember using the pass through in CSP, but I forgot what it is supposed to be.

6:45pm. I just realized something. Tumble and dolly based on where you grab and zoom as opposed to the object center does make sense. In this case it does make it easier to move around.

Let me stop here for lunch. I am overwhelmed by all the options right now. Even though I have the desk in front of me, I am not sure how to begin work on the intricate noise pattern.

6:55pm. When I am in Substance Painter, it feels very similar to Clip Studio Paint.

Right now, it is not that I can't do these patterns, but the problem is that there too many ways of doing them. I can't do them exactly, so I need to pick a simplified way of doing it. There are 100s of brushes to choose from, not to mention textures and materials.

It will take me a while to overcome the inertia.

7:10pm. I really don't feel like doing this anymore for the day. It is torture to have Pt open at this hour. Yesterday I might have gone till past 12am, but that was only because I obsessed about a problem that needed solving. Right now I am still fatigued from getting up too early, and do not have anything immediate I want to do. I just need to study and explore. That kind of play is best done when fresh.

7:55pm. Needed that break. At this point I am really lacking in Pt fundamentals. I need to make up my mind to gradually get into it.

8pm. I'll let it come to me, just not when I am high on stress like this. Sure it might have been better to not spend the most productive part of day messing with UVs, but an issue is finished when you feel satisfied deep down, not when you tell yourself its over. Similarly, I think the main reason why I got up this early is because I really did feel the need to readjust my schedule even if it meant a rough day.

Let me close here. Time for rest and indulgence.

I am going learn Pt and to finish the desk before long. The rest of the props should go a lot faster after that."