"10:50am. Got up 20m ago. Let me do my morning reading. It is good that I am slowly adjusting my schedule.
11am. I feel like chilling for a while longer. Who feels like starting? Slacking is food for the soul.

11:25am. Let me start. There is one thing I've realized - textures are awful. They are the worst idea in 3d, but the ability to do it all procedurally is not there. I am thinking of studying Designer for a bit just so I can get a grasp on what it is about.

11:35am. That procedural texturing course I got, let me watch some of it. I'll download Ds while that is going on.

11:50am. It seems the course will explain what the texture coordinates in Blender are. Nice. The Houdini + Clarisse course I got previously was mediocre, but this seems like it could be good.

11:55am. This explanation of coordinate spaces is so good. When I was starting to study shading, this is the kind of course that I needed. It gets points just for that.

12:20pm. This course is not something you can watch casually. It is quite technical. I'd need to follow it along in order to grasp it. Let me watch a Designer tutorial. I want to see how it works.

https://substance3d.adobe.com/tutorials/courses/First-Steps-with-Substance-3D-Designer/youtube-VyFgpitTsYg

Let me watch some of these.

12:25pm. Hmmm, can you really not do your texturing in Designer. It has a visualizer, so I do not see any problem.

https://youtu.be/VyFgpitTsYg?t=198

Oh, this has the radial patterns as expected

I have to admit at this point, this way of chaining parameterized nodes is quite something.

The closest analogue in functional programming is using partially applied functions (combinators) and those are quite effective in their domain, such as parsing.

12:40pm. Let me watch the first part and then I'll have breakfast. I am starting to get into it.

12:45pm. I wonder if it is possible to paint directly in Designer?

12:50pm. I hate how RapidGator seems to be designed just to annoy me. AlfaFile might be slower, but it does not bother me what I forget about the captcha for half an hour.

https://youtu.be/VyFgpitTsYg?t=693

I am here so far. This is powerful functionality. There is no reason why I can't use both Designer and Painter.

When I am strapped for time I feel like limiting myself, but when I am true to my desires I spend time in exploration.

Yeah, it is not worth it to force myself to do Heaven's Key. I am fine with going towards it, but if I really need money I'll just get a programming job.

https://youtu.be/VyFgpitTsYg?t=818

This is nice.

Let me watch part 2 as well, after that I'll take a break.

https://youtu.be/LJFgZvNGN9U?t=315

Here is the warp node. This is what I've been wanting all along.

https://youtu.be/LJFgZvNGN9U?t=445

Directional warp. Exactly what I'd need. Maybe the guy who said that Painter without Designer is a car without wheels was right.

Just who am I? Of course I should be aiming to master both. I'd want to add Mari into the mix even if it takes me much longer.

That scratch floor texture I did in Houdini? I could have done it in a single stroke in Painter.

I'll want to figure out how to import the V-Ray asset library into Substance Painter.

https://youtu.be/LJFgZvNGN9U?t=684

It is that easy to interop between the programs? Nice.

I might as well watch part 3. I am into it.

https://youtu.be/22fZAM6r6Hg?t=305

There is a safe transform it seems.

1:30pm. Watched the vids.

1:30pm. Since things are the way they are, why not learn both Painter and Designer at the same time? Now if only this would download. Right now I am waiting for the Gator timeout to run out.  I'll go back to the other basic Painter tutorial and go through it. By the time I am done with it, Designer should have finished downloading.

It is fine if I immerse myself in this. I mean apart from texturing and shading what else is there before I can consider myself to be proficient at 3d art? Not much. I am already decently along the learning curve, I just need to seal the deal. So what if it take me a couple of weeks to get through all of this?

It is nothing.

While I was programming, did I rush the creation of Spiral? No. And neither will I rush this. There is no reason for me to not create my own materials when I need them. It is not an activity I should fear.

I will master it all and get to the highest level in 3d. After I do that I'll be able to create whatever I want."