"10pm. Let me put in a bit extra. I really want to watch that shape manipulation video.
https://youtu.be/11GlyrYNbnA?t=187

What is gradient dynamic doing here. This is arcane work.

I guess Substance Designer will carve its own legend as an art software.

https://substance3d.adobe.com/documentation/sddoc/gradient-dynamic-172825217.html

> The Gradient (Dynamic) or Dynamic Gradient node remaps a grayscale input to a new grayscale or color ramp.

> It serves as a slight alternative to the Gradient Node, but unlike the Gradient node, Gradient color keys are not defined internally, but come from an external input. This mainly allows to avoid the problem where parameters cannot be exposed, as the parameters for color are moved outside of the node. This is what makes it "dynamic".

How does this work?

https://youtu.be/o8AtdW1oFC8?t=118

This really explains a lot. Still what do the different columns represent in the gradient input?

https://youtu.be/o8AtdW1oFC8?t=138

This tiling is exactly how I meant to do the wood texture. Well rather than tiling, I'd mirror the gradients so it is smooth on both sides.

https://substance3d.adobe.com/documentation/sddoc/gradient-dynamic-172825217.html

Ah, I see. You select the column or the row based on the third parameter which is the input position. I am too tired to think properly it seems.

https://substance3d.adobe.com/documentation/sddoc/color-burn-159450235.html
> Performs a Color Burn blend between Foreground and Background. Mathematically the formula is 1 - (1-Background) / Foreground.

So that is what the formula is!

https://substance3d.adobe.com/documentation/sddoc/color-dodge-159450239.html
> Performs a Color Dodge blend. Mathematically the formula is Background / (1-Foreground).

I've been trying to figure out what the blending modes do mathematically since forever. I finally found it here. Google was completely useless when it came to explaining them. It just does the equivalent of linking me to articles that say - this blend mode makes things darker, or this blend mode makes things lighter. What was I supposed to do with that?

https://substance3d.adobe.com/tutorials/courses/Designer-Quicktips/youtube-hIpTq8vZB9Q

Let me watch basic shape manipulation and then I am really done.

https://youtu.be/hIpTq8vZB9Q?t=94

Oh a direction blur is exactly what I need for the radial pattern. I want it to get elongated in the x axis the further it gets from the center.

10:30pm. https://80.lv/articles/discovering-the-curve-node-in-substance-designer/

The curve node is exactly what Shrimp showed me in his short video. I could use it to shape the circle manually.

Seeing his do his tricks really had a powerful impression on me. I really gained a lot just from watching those 10m of doing his magic.

Working in Designer is basically a whole new artform that has been invented in the 21st century. It is different from programming, the brain probably doing something closer to working in Designer than with regular programming.

This is what I've been missing. Doing the kinds of things done here is not something I would ever imagine doing in a regular language.

10:40pm. https://80.lv/articles/discovering-the-curve-node-in-substance-designer/

I'll leave this article for later. I can't be bothered to read it right now. Let me get back to Ayakashi."