"10:25am. Let me finish Frieren and I'll watch the talk by Schmidhuber. I feel groggy and am not in the mood for making mud right off the bat.
https://youtu.be/2GgGVdkq2bU
Meta-Learning Machines in a Single Lifelong Trial

Schmidhuber is getting into the Youtuber game.

https://youtu.be/2GgGVdkq2bU?t=200

Maybe he is right. I would bet on genetic programming, but that is because I don't know how to make use of RNNs to do what he is talking about.

https://youtu.be/2GgGVdkq2bU?t=591

Hmmm...maybe I should try thinking about this seriously. Placing checkpoints makes sense, but they'd have to be rough and spaced out in time. Even if this worked, it is not a problem, the problem is having an algorithm that can learn representations in a stable manner while doing self improvement.

10:50am. Maybe using checkpoints would be more principled than using a population based approach. But the later could have diversity benefits.

And what does it matter either way, trying to stich together code would be hard anyway.

11:05am. He is just rehashing his old stuff here. Nothing new to talk about. Forget it. Let me chill a bit more and then I will get started.

11:30am. Time for some mud action.

Let me get on with it.

https://substance3d.adobe.com/tutorials/courses/Getting-Started-with-Substance-3D-Designer/youtube-y8q6-tgQjZc
Getting Started With Substance 3d Designer

https://youtu.be/y8q6-tgQjZc?t=60

Oh wow, you can actually bring in substance materials directly into Maya. Though in that Blender course, Shrimp did say that blurs would be impossible. So how would that work?

11:45am. Based on the intro section, I should gain a lot once I go through this course. In the tree bark one he just flies over the nodes and does not explain much of his thought process. But if I could understand the things in this course, I'll be able to do vastly more than the knot patterns.

https://youtu.be/k6MrHOGBy84?t=303

So I can link my own 3d shapes like this.

https://youtu.be/LvvsuKhc3wc?t=555

The video explaining to me how to do the materials is actually super useful.

1pm. https://substance3d.adobe.com/tutorials/courses/Getting-Started-with-Substance-3D-Designer/youtube-wGBsKB5sFMw

I was really starting to get into it and lost track of the time. Let me pause here.

https://mangadex.org/title/4c86f38a-cead-4575-a9c3-acf7261a58ff/meikyuu-black-company

It was a complete accident that I saw this being uploaded on Dex's feed. I never expected translations for this given that its licensed. Nice. Things are looking good. Also I remembered that I am yet to finish UQ Holder. Let me focus on that during the break.
"