"10:05am. I slept much better than yesterday and even had time to chill. Let me start. With the Wood Plank 1.5h course.
https://substance3d.adobe.com/tutorials/courses/Creating-Old-Wood-Planks-in-Substance-3D-Designer/youtube-vLz8ZyduQi0

Oh, I was wrong. This is 3:45h long.

https://substance3d.adobe.com/tutorials/courses/Designer-Quicktips/youtube-Xv-CfrMtS_M
https://substance3d.adobe.com/tutorials/courses/Getting-Started-with-MaterialX/youtube-bduf3TWmeMY

What is this about MaterialX.

Ok, let me just focus on studying today. I'll learn what I need and go back to Painter tomorrow.

https://youtu.be/S8P4QAH2K1U
Wood Grain

There is a tutorial on exactly what I wanted to do. This is a bit different than how I imagined it doing though. But there are plenty of ways of doing grains as expected.

10:35am. Let me start with the Wood Plank course. I'll just watch it for a while instead of following directly along as actually practicing would take too long.

11am. I saw this in one of the quicktips. If you double click on the empty space, you'll see random seed on the graph properties. That answers the question of how to randomize the seeds globally.

11:05am. Had to take a break, let me resume.

https://youtu.be/IcwDAoegKsg?t=581

His techniques are interesting. This is not something I would have figured out on my own. I guess some things really have to taught to me.

Yes, sure, maybe if I sat down and mucked around with these nodes at random for months I could have figured it out. But that is just in theory. Every loser thinks like that. I could have figure out AI using that same technique.

No. The best way to learn these techniques is to steal them. Let some other fool play around figure out the useful things. I am happy to get them from him.

https://youtu.be/frtvwbbtfOk
How to create tileable/seamless textures in Substance Painter

Let me pause so I can watch this 2m video. It is strange that Pt does not have an inbuilt plane model.

https://youtu.be/frtvwbbtfOk?t=46

Ah, right. Yesterday I had the idea of setting a brush to UV space. That would make it wrap around the edge of a plane. Wouldn't that work? Why do this then? Maybe it would not draw past the edge of a plane unless I shrunk it.

https://youtu.be/frtvwbbtfOk?t=74

He is doing that in fact. Seeing this example is giving me insight into how to do ornamental patterns.

Let me get back to the wood course.

What is Emboss?

https://youtu.be/jMs1U7G_LHQ
Emboss Node [Atomic Nodes] - Substance Designer (showcase)

That wood course is useful for teaching me things like this. I want to fill just a bit more of the gaps before I move to Painter for good.

Hmmm, it is an interesting effect. I get it now.

11:40am. I thought that Houdini would be good for heightfields, but this is quite something as well. Painter would be equally good as well.

What he is doing here at the start with just mixing different noises is something that could be done in Painter.

https://youtu.be/N4aNIDJ84nM?t=479

This part cofuses me. I am not really seeing the tiling break.

But that might just be because my focus is low. As expected, if you are just listeting to a talk, you are not into it as much if you are following along.

12:15pm. Ah, I should have beveled the rocks yesterday instead of bluring them. I need to remember to try that out in the future.

https://substance3d.adobe.com/tutorials/courses/Creating-Old-Wood-Planks-in-Substance-3D-Designer/youtube-we4lMxU94CA

I won't watch this. I am done with studying Designer. Good job me.

Now it is time to get real. I've built up my foundation in Painter enough. What I just need to do is use it and I will reach the level of expertise in 3d that I've been seeking.

I've got Blender and Houdini skills. It is time to start putting them together.

Now let me chill here for a while.

I am not going to rush this. I'll take it slowly today. I've worked so hard to get over the hurdles in 3d. The shading parts were rocky for a while, but they won't be getting in my way any longer.

2:25pm. Let me close Dendro. It is really hard to pry myself off, as I was starting to get into it.

2:30pm. What is hardship? Nothing.

Being able to say it is nothing is what separates an adult from a child.

The hardest thing in the world is to make a single single step forward. It is an adult thing to keep doing it regardless.

I did not change from my time in high school. The only thing that changed is that my true path is clearer. I cannot go faster, but I know where to go now.

2:35pm. These art skills are nothing. Six months of practice are nothing.

I've been doing it every day, and I will do it every day until I find the power that I seek.

One day the power of the machines will be mine. But today, let deal with the desk.

2:40pm. Just when is good enough, good enough? Right now it should be. I've put in enough work to consider myself an adult.

If I had a better background, I wouldn't have failed something like trading.
if I had a better background, I would not have lacked resources for the AI chips.

But in the long run those things do not matter. A 1000 years from now, the hurdles I've been facing now will just be a blimp on the timeline. It is a cheap price to pay to attain the priestly skill set.

The fact that I could go through the programming hurldes is why I can so fearlessly go into art.

2:45pm. i am not going to be doing my own props forever. I'll do some special cases like now, but afterwards I'll get a Alfafile or RapidGator acccount and get some proper kits. For models, I can just do them myself, but it is not like I can't get finished ones and adjust them.

2:50pm. With effort, I can succeed with art. So let me start.

4:40pm. Let me take a break. It is not going well.

5:35pm. Sigh, I am a fool. I am doing everything the opposite way around.

8:15pm. Fuck! Aren't the filters for Substance Painter documented anywhere? Why does the directional warp not work for me?

9:40pm. Oh, it actually crashed.

12:40pm. Fuck. I did not finish it despite being at it all this time.

12:55pm. Stop me, stop! I'll leave it for tomorrow.

I think I made a mistake trying to do this in Painter. I should have done the texture in Designer.

1:50am. Let me go to bed. I'll try it in Designer tomorrow. I'll talk about it more when I rest up."